auth api

+++ app/Http/Controllers/Api/AuthController.php
@@ -1,14 +1,14 @@
 <?php
 
 namespace App\Http\Controllers\Api;
-use App\Models\Post;
+use App\Models\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\PostResource;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Storage;
 
-class PostController extends Controller
+class AuthController extends Controller
 {
     /**
      * index
@@ -18,7 +18,7 @@ class PostController extends Controller
     public function index()
     {
         //get all posts
-        $posts = Post::latest()->paginate(5);
+        $posts = User::latest()->paginate(5);
 
         //return collection of posts as a resource
         return new PostResource(true, 'List Data Posts', $posts);
@@ -34,9 +34,9 @@ class PostController extends Controller
     {
         //define validation rules
         $validator = Validator::make($request->all(), [
-            'image'     => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-            'title'     => 'required',
-            'content'   => 'required',
+            'name' => 'required',
+            'email' => 'required|email',
+            'password' => 'required'
         ]);
 
         //check if validation fails
@@ -44,15 +44,11 @@ class PostController extends Controller
             return response()->json($validator->errors(), 422);
         }
 
-        //upload image
-        $image = $request->file('image');
-        $image->storeAs('public/posts', $image->hashName());
-
         //create post
-        $post = Post::create([
-            'image'     => $image->hashName(),
-            'title'     => $request->title,
-            'content'   => $request->content,
+        $post = User::create([
+            'name'     => $request->name,
+            'email'   => $request->email,
+            'password'   => $request->password,
         ]);
 
         //return response
@@ -68,7 +64,7 @@ class PostController extends Controller
     public function show($id)
     {
         //find post by ID
-        $post = Post::find($id);
+        $post = User::find($id);
 
         //return single post as a resource
         return new PostResource(true, 'Detail Data Post!', $post);
@@ -85,8 +81,9 @@ class PostController extends Controller
     {
         //define validation rules
         $validator = Validator::make($request->all(), [
-            'title'     => 'required',
-            'content'   => 'required',
+            'name' => 'required',
+            'email' => 'required|email',
+            'password' => 'required'
         ]);
 
         //check if validation fails
@@ -95,32 +92,14 @@ class PostController extends Controller
         }
 
         //find post by ID
-        $post = Post::find($id);
+        $post = User::find($id);
 
-        //check if image is not empty
-        if ($request->hasFile('image')) {
-
-            //upload image
-            $image = $request->file('image');
-            $image->storeAs('public/posts', $image->hashName());
-
-            //delete old image
-            Storage::delete('public/posts/' . basename($post->image));
-
-            //update post with new image
-            $post->update([
-                'image'     => $image->hashName(),
-                'title'     => $request->title,
-                'content'   => $request->content,
-            ]);
-        } else {
-
-            //update post without image
-            $post->update([
-                'title'     => $request->title,
-                'content'   => $request->content,
-            ]);
-        }
+        //update post without image
+        $post->update([
+            'name'     => $request->name,
+            'email'   => $request->email,
+            'password'   => $request->password,
+        ]);
 
         //return response
         return new PostResource(true, 'Data Post Berhasil Diubah!', $post);
@@ -135,15 +114,23 @@ class PostController extends Controller
     public function destroy($id)
     {
         //find post by ID
-        $post = Post::find($id);
+        $post = User::find($id);
 
-        //delete image
-        Storage::delete('public/posts/'.basename($post->image));
+        //check if post exists
+        if (!$post) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Data Post Tidak Ditemukan!',
+            ], 404);
+        }
 
         //delete post
         $post->delete();
 
         //return response
-        return new PostResource(true, 'Data Post Berhasil Dihapus!', null);
+        return response()->json([
+            'success' => true,
+            'message' => 'Data Post Berhasil Dihapus!',
+        ], 200);
     }
 }
